+++ contracts/commands/admin/Annotate.sol
@@ -21,10 +21,10 @@ abstract contract Annotate is AdminBase {
     /// @return Empty transaction stream.
     function annotate(
         bytes32 account,
-        bytes calldata,
+        bytes calldata state,
         bytes calldata input
     ) external onlyAdmin(account) returns (bytes memory, bytes memory) {
-        Execution memory exec = openInput(input, descriptor, 0);
+        Execution memory exec = openCommand(state, input, descriptor, 0);
 
         while (exec.more()) {
             (uint entity, bytes calldata data) = exec.unpackAnnotation(Lanes.Input);