+++ contracts/commands/admin/Appoint.sol
@@ -22,10 +22,10 @@ abstract contract Appoint is GuardianAccess, AdminBase {
     /// @return Empty transaction stream.
     function appoint(
         bytes32 account,
-        bytes calldata,
+        bytes calldata state,
         bytes calldata input
     ) external onlyAdmin(account) returns (bytes memory, bytes memory) {
-        Execution memory exec = openInput(input, descriptor, 0);
+        Execution memory exec = openCommand(state, input, descriptor, 0);
 
         while (exec.more()) {
             bytes32 guardian = exec.unpackAccount(Lanes.Input);